MOD / 3-2-3. 유기견 수정 / 구현완료

+++ springboot/src/main/java/com/petmeeting/springboot/domain/Users.java
@@ -53,7 +53,7 @@ public abstract class Users {
 
     @PrePersist
     public void prePersist() {
-        this.joinDate = joinDate == null ? (int) System.currentTimeMillis() / 1000 : joinDate;
+        this.joinDate = joinDate == null ? System.currentTimeMillis() / 1000 : joinDate;
     }
 
     @OneToMany(mappedBy = "user", fetch = FetchType.LAZY)
